refactor: extract section toggle into separate view
